+++ app/Models/Article.php
@@ -2,11 +2,7 @@
 
 namespace App\Models;
 
-use App\Models\Concerns\HasComments;
-use App\Models\Concerns\HasLikes;
 use App\Models\Presenters\ArticlePresenter;
-use App\Support\Markdown\Markdown;
-use Carbon\CarbonInterval;
 use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Support\Str;
 use Illuminate\Database\Eloquent\Model;
@@ -61,5 +57,17 @@ class Article extends Model implements Feedable
     public function getFeedResults()
     {
         return static::query()->published()->free()->latest('published_at')->get();
+    }
+
+    public function toFeedItem()
+    {
+        return FeedItem::create([
+            'id' => $this->id,
+            'title' => $this->title,
+            'summary' => $this->excerpt,
+            'updated' => $this->updated_at,
+            'link' => route('articles.show', $this),
+            'author' => 'Ryan Chandler',
+        ]);
     }
 }
