+++ composer.json
@@ -18,6 +18,7 @@
         "laravel/socialite": "^4.3",
         "laravel/tinker": "^2.0",
         "league/commonmark": "^1.4",
+        "spatie/browsershot": "^3.40",
         "spatie/commonmark-highlighter": "^2.1",
         "spatie/laravel-feed": "^2.7",
         "spatie/laravel-sitemap": "^5.8"
