+++ package.json
@@ -16,5 +16,8 @@
         "postcss-import": "^12.0.1",
         "tailwindcss": "^1.8",
         "tiny-glob": "^0.2.6"
+    },
+    "dependencies": {
+        "puppeteer": "^5.3.1"
     }
 }
